+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -9,6 +9,7 @@ import createPlugins from './createPlugins';
 import { createKeyBindingFn, initPluginKeyBindings } from './keyBindings';
 import handleKeyCommand from './handleKeyCommand';
 import handleReturnCommand from './handleReturnCommand';
+import handlePastedText from './handlePastedText';
 import blockStyleFn from './blockStyleFn';
 import { combineStyleFns } from './combineStyleFns';
 import { getStaticTextToolbarId } from './Toolbars/toolbar-id';
@@ -21,7 +22,6 @@ import {
   getFocusedBlockKey,
   createCalcContentDiff,
   getPostContentSummary,
-  Modifier,
   getBlockType,
   COMMANDS,
   MODIFIERS,
@@ -37,8 +37,6 @@ import {
 import styles from '../../statics/styles/rich-content-editor.scss';
 import draftStyles from '../../statics/styles/draft.rtlignore.scss';
 import 'wix-rich-content-common/dist/statics/styles/draftDefault.rtlignore.scss';
-import { convertFromHTML as draftConvertFromHtml } from 'draft-convert';
-import { pastedContentConfig, clearUnnecessaryInlineStyles } from './utils/pastedContentUtil';
 import { deprecateHelpers } from 'wix-rich-content-common/dist/lib/deprecateHelpers.cjs.js';
 
 class RichContentEditor extends Component {
@@ -302,32 +300,6 @@ class RichContentEditor extends Component {
     this.props.onChange?.(editorState);
   };
 
-  handlePastedText = (text, html, editorState) => {
-    const { handlePastedText } = this.props;
-    if (handlePastedText) {
-      return handlePastedText(text, html, editorState);
-    }
-    if (html) {
-      const htmlContentState = draftConvertFromHtml(pastedContentConfig)(html);
-      const contentState = Modifier.replaceWithFragment(
-        editorState.getCurrentContent(),
-        editorState.getSelection(),
-        htmlContentState.getBlockMap()
-      );
-      const newEditorState = EditorState.push(editorState, contentState, 'insert-fragment');
-      const newContentState = clearUnnecessaryInlineStyles(contentState);
-      const resultEditorState = EditorState.set(newEditorState, {
-        currentContent: newContentState,
-        selection: newEditorState.getSelection(),
-      });
-
-      this.updateEditorState(resultEditorState);
-      return 'handled';
-    } else {
-      return false;
-    }
-  };
-
   handleTabCommand = () => {
     if (this.getToolbars().TextToolbar) {
       const staticToolbarButton = this.findFocusableChildForElement(
@@ -337,6 +309,17 @@ class RichContentEditor extends Component {
     } else {
       this.editor.blur();
     }
+  };
+
+  handlePastedText = (text, html, editorState) => {
+    if (this.props.handlePastedText) {
+      return this.props.handlePastedText(text, html, editorState);
+    }
+
+    const resultEditorState = handlePastedText(text, html, editorState);
+    this.updateEditorState(resultEditorState);
+
+    return 'handled';
   };
 
   getCustomCommandHandlers = () => ({