+++ packages/editor/web/test/TestData/pasted-data.js
@@ -1,0 +1,92 @@
+export const raw = {
+  blocks: [
+    {
+      key: 'ap77n',
+      text: '',
+      type: 'unstyled',
+      depth: 0,
+      inlineStyleRanges: [],
+      entityRanges: [],
+      data: {},
+    },
+    {
+      key: '6hc2b',
+      text: ' ',
+      type: 'atomic',
+      depth: 0,
+      inlineStyleRanges: [],
+      entityRanges: [
+        {
+          offset: 0,
+          length: 1,
+          key: 0,
+        },
+      ],
+      data: {},
+    },
+    {
+      key: 'cqrco',
+      text: '',
+      type: 'unstyled',
+      depth: 0,
+      inlineStyleRanges: [],
+      entityRanges: [],
+      data: {},
+    },
+  ],
+  entityMap: {
+    '0': {
+      type: 'wix-draft-plugin-image',
+      mutability: 'IMMUTABLE',
+      data: {
+        config: {
+          alignment: 'center',
+          size: 'content',
+          showTitle: true,
+          showDescription: true,
+        },
+        src: {
+          id: '17ef104520dd9a0711f1cc900d970b79',
+          original_file_name: '8bb438_277609c27d9346c38d480b6da8237820.jpg',
+          file_name: '8bb438_277609c27d9346c38d480b6da8237820.jpg',
+          width: 1920,
+          height: 1280,
+        },
+      },
+    },
+  },
+  VERSION: '7.10.5',
+};
+
+export const expectedRaw = {
+  blocks: [
+    {
+      key: 'ap77n',
+      text: '',
+      type: 'unstyled',
+      depth: 0,
+      inlineStyleRanges: [],
+      entityRanges: [],
+      data: {},
+    },
+    {
+      key: '6hc2b',
+      text: '3000',
+      type: 'unstyled',
+      depth: 0,
+      inlineStyleRanges: [],
+      entityRanges: [],
+      data: {},
+    },
+    {
+      key: 'cqrco',
+      text: '',
+      type: 'unstyled',
+      depth: 0,
+      inlineStyleRanges: [],
+      entityRanges: [],
+      data: {},
+    },
+  ],
+  entityMap: {},
+};